fix リストを分離
