+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "convex-management",
-  "version": "1.0.0",
+  "version": "6.1.0",
   "description": "",
   "main": "index.js",
   "scripts": {
