+++ src/convex/report/index.ts
@@ -1,21 +1,21 @@
 import * as Discord from 'discord.js'
 import Option from 'type-of-option'
 import Settings from 'const-settings'
+import * as update from './update'
+import * as list from './list'
 import * as current from '../../io/current'
-import * as damageList from '../../io/damageList'
 import * as status from '../../io/status'
 import * as declare from '../declare'
 import * as declareList from '../declare/list'
 import * as declareStatus from '../declare/status'
 import * as over from '../over'
 import * as cancel from '../plan/delete'
-import * as list from '../plan/list'
+import * as planList from '../plan/list'
 import * as role from '../role'
 import * as situation from '../situation'
 import * as limit from '../time/limit'
-import * as update from './update'
 import * as util from '../../util'
-import {AtoE, Current, Member} from '../../util/type'
+import {AtoE, Member} from '../../util/type'
 
 /**
  * 凸報告の管理を行う
@@ -81,7 +81,7 @@ export const Convex = async (msg: Discord.Message): Promise<Option<string>> => {
   let state = await current.Fetch()
   const overMsgs = await over.GetAllUserMsg(member_2.id)
 
-  content = await peportConfirm(members, member_2, state, alpha, overMsgs, content, msg)
+  content = await list.Reply(members, member_2, state, alpha, overMsgs, content, msg)
 
   // @が入っている場合、HPの変更
   if (/@\d/.test(content)) {
@@ -91,7 +91,7 @@ export const Convex = async (msg: Discord.Message): Promise<Option<string>> => {
   // `;`が入っている場合は凸予定を取り消さない
   if (!/;/i.test(content)) {
     cancel.Remove(alpha, msg.author.id)
-    list.SituationEdit()
+    planList.SituationEdit()
   }
 
   overDelete(member_2, carry, overMsgs)
@@ -124,60 +124,6 @@ const threeConvexProcess = async (member: Member, msg: Discord.Message): Promise
   await msg.reply(`残凸数: 0、持越数: 0\n\`${n}\`人目の3凸終了よ！`)
 
   return true
-}
-
-/**
- * 残りの凸状況を報告する
- * @param members メンバー全員の状態
- * @param member メンバーの状態
- * @param state 現在の状況
- * @param alpha ボスの番号
- * @param overMsgs メンバーの持越状況のメッセージ一覧
- * @param content 凸報告のメッセージ
- * @param msg DiscordからのMessage
- */
-const peportConfirm = async (
-  members: Member[],
-  member: Member,
-  state: Current,
-  alpha: AtoE,
-  overMsgs: Discord.Message[],
-  content: string,
-  msg: Discord.Message
-): Promise<string> => {
-  const damages = await damageList.FetchBoss(alpha)
-  damages
-  overMsgs
-
-  const boss = state[alpha]
-
-  // 凸報告のメッセージからHPを取得
-  let hp: Option<number | string> = content
-    .replace(/^.*@/g, '')
-    .trim()
-    .replace(/\s.*$/g, '')
-    .match(/\d*/)
-    ?.map(e => e)
-    .first()
-  hp = hp === '' || hp === undefined ? boss.hp : hp
-  const maxHP = Settings.STAGE[state.stage].HP[alpha]
-
-  // 何人3凸終了しているか確認
-  const endN = members.filter(s => s.end).length
-
-  // channel.send(`<@!${member.id}> <#${Settings.CHANNEL_ID.CARRYOVER_SITUATION}> を整理してね`)
-
-  const text = [
-    '```m',
-    `${boss.lap}周目 ${boss.name} ${hp}/${maxHP}`,
-    `残凸数: ${member.convex}、持越数: ${member.over}`,
-    member.end ? `${endN}人目の3凸終了よ！` : '',
-    '```',
-  ].join('\n')
-
-  await msg.reply(text)
-
-  return content
 }
 
 /**
